requirements.txt
aiotg
ASF_IPC>=2.1.0
